--- Ebay/src/test/resources/EbayTestRunner.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-
-<suite name="Ebay Automation Script " parallel="false">
-
-    <parameter name="useCloudEnv" value="false"/>
-    <parameter name="cloudEnvName" value="browserstack"/>
-    <parameter name="os" value="OS X"/>
-    <parameter name="os_version" value="Sierra"/>
-    <parameter name="browserName" value="chrome"/>
-    <parameter name="browserVersion" value="79.0"/>
-    <parameter name="url" value="https://www.ebay.com"/>
-
-    <test name="Ebay Home">
-        <classes>
-            <class name="home.EbayHome"/>
-        </classes>
-    </test>
-
-</suite>
